refactor(api): Extract StatsService from StatsController
