+++ app/Http/Controllers/API/v1/StatsController.php
@@ -3,46 +3,20 @@
 namespace App\Http\Controllers\API\v1;
 
 use App\Http\Controllers\API\ApiController;
-use App\Models\Transaction;
 use App\Models\Wallet;
-use App\Services\ExchangeRateService;
+use App\Services\StatsService;
 use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
 use OpenApi\Attributes as OA;
-use RuntimeException;
 
 #[OA\Tag(name: 'Statistics', description: 'Financial statistics and analytics')]
 class StatsController extends ApiController
 {
-    protected ExchangeRateService $exchangeRateService;
-
-    public function __construct(ExchangeRateService $exchangeRateService)
-    {
-        $this->exchangeRateService = $exchangeRateService;
-    }
-
-    /**
-     * Convert amount to target currency, throwing exception on failure.
-     *
-     * @throws \RuntimeException When currency conversion fails
-     */
-    private function convertCurrency(float $amount, string $fromCurrency, string $targetCurrency, $user): float
-    {
-        if ($fromCurrency === $targetCurrency) {
-            return $amount;
-        }
-
-        $converted = $this->exchangeRateService->convert($amount, $fromCurrency, $targetCurrency, $user);
-
-        if ($converted === null) {
-            throw new RuntimeException(
-                "Failed to convert {$fromCurrency} to {$targetCurrency}. Exchange rate unavailable."
-            );
-        }
-
-        return $converted;
+    public function __construct(
+        protected StatsService $statsService
+    ) {
     }
 
     #[OA\Get(
@@ -117,1116 +91,95 @@ class StatsController extends ApiController
             ),
         ]
     )]
-    /**
-     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
-     */
     public function index(Request $request): JsonResponse
     {
         $user = $request->user();
 
-        // Set default date range
-        $endDate = Carbon::now()->endOfDay();
-        $startDate = Carbon::now()->subDays(30)->startOfDay();
+        [$startDate, $endDate] = $this->resolveDateRange($request);
 
-        // Handle preset periods (overrides start_date/end_date)
-        if ($request->has('preset')) {
-            $preset = $request->input('preset');
-            $endDate = Carbon::now()->endOfDay();
-
-            switch ($preset) {
-                case 'all_time':
-                    $startDate = Carbon::parse('2000-01-01')->startOfDay();
-                    break;
-                case 'current_week':
-                    $startDate = Carbon::now()->startOfWeek()->startOfDay();
-                    break;
-                case 'current_month':
-                    $startDate = Carbon::now()->startOfMonth()->startOfDay();
-                    break;
-                case 'last_3_months':
-                    $startDate = Carbon::now()->subMonths(3)->startOfDay();
-                    break;
-                default:
-                    $startDate = Carbon::now()->subDays(30)->startOfDay();
-            }
-        } else {
-            // Apply date filters if provided
-            if ($request->has('start_date')) {
-                $startDate = Carbon::parse($request->input('start_date'))->startOfDay();
-            }
-
-            if ($request->has('end_date')) {
-                $endDate = Carbon::parse($request->input('end_date'))->endOfDay();
-            }
-        }
-
-        // Get wallet IDs filter
-        $walletIds = [];
-        if ($request->has('wallet_ids')) {
-            $walletIds = explode(',', $request->input('wallet_ids'));
-            $walletIds = array_filter(array_map('intval', $walletIds));
-
-            // Validate that all provided wallet IDs exist and belong to the user
-            $validWalletIds = Wallet::where('user_id', $user->id)
-                ->whereIn('id', $walletIds)
-                ->pluck('id')
-                ->toArray();
-
-            // Check if all provided wallet IDs are valid
-            $invalidWalletIds = array_diff($walletIds, $validWalletIds);
-            if (! empty($invalidWalletIds)) {
-                return response()->json([
-                    'message' => __('One or more wallet IDs are invalid or do not belong to the user.'),
-                    'invalid_wallet_ids' => array_values($invalidWalletIds),
-                ], 422);
-            }
+        $walletIds = $this->resolveWalletIds($request, $user);
+        if ($walletIds instanceof JsonResponse) {
+            return $walletIds;
         }
 
         $defaultCurrency = $user->getConfigValue('default-currency') ?? 'USD';
+        $period = $request->input('period', 'month');
 
-        // Generate cache key based on user, dates, and filters
-        $cacheKey = $this->generateCacheKey(
+        $cacheKey = StatsService::generateCacheKey(
             $user->id,
             $startDate,
             $endDate,
             $walletIds,
-            $request->input('period', 'month')
+            $period
         );
 
         $data = Cache::remember(
             $cacheKey,
             now()->addMinutes(5),
-            function () use (
+            fn () => $this->statsService->compute(
                 $user,
                 $startDate,
                 $endDate,
                 $walletIds,
-                $request,
+                $period,
                 $defaultCurrency
-            ) {
-                $periodTotals = $this->getPeriodTotals($user, $startDate, $endDate, $walletIds, $defaultCurrency);
-
-                $overview = [
-                    'total_balance' => $this->getTotalBalance($user, $walletIds, $defaultCurrency),
-                    'net_worth' => $this->getNetWorth($user, $walletIds, $defaultCurrency),
-                    'total_income' => $periodTotals['income'],
-                    'total_expenses' => $periodTotals['expense'],
-                    'net_cash_flow' => $periodTotals['income'] - $periodTotals['expense'],
-                    'avg_monthly_income' => $this->getAverageMonthlyAmount(
-                        $user,
-                        'income',
-                        $walletIds,
-                        $startDate,
-                        $endDate,
-                        $defaultCurrency
-                    ),
-                    'avg_monthly_expenses' => $this->getAverageMonthlyAmount(
-                        $user,
-                        'expense',
-                        $walletIds,
-                        $startDate,
-                        $endDate,
-                        $defaultCurrency
-                    ),
-                    'savings_rate' => 0,
-                ];
-
-                if ($overview['total_income'] > 0) {
-                    $overview['savings_rate'] =
-                    (($overview['total_income'] - $overview['total_expenses']) / $overview['total_income']) * 100;
-                }
-
-                $activityMetrics = $this->getActivityMetrics($user, $startDate, $endDate, $walletIds);
-
-                return [
-                    'currency' => $defaultCurrency,
-                    'overview' => $overview,
-                    'activity' => $activityMetrics,
-                    'comparisons' => $this->getComparisons($user, $startDate, $endDate, $walletIds, $defaultCurrency),
-                    'top_categories' => $this->getTopCategories(
-                        $user,
-                        $startDate,
-                        $endDate,
-                        $walletIds,
-                        $defaultCurrency
-                    ),
-                    'largest_transactions' => $this->getLargestTransactions(
-                        $user,
-                        $startDate,
-                        $endDate,
-                        $walletIds,
-                        $defaultCurrency
-                    ),
-                    'spending_trends' => $this->getSpendingTrends(
-                        $user,
-                        $startDate,
-                        $endDate,
-                        $request->input('period', 'month'),
-                        $walletIds,
-                        $defaultCurrency
-                    ),
-                    'category_distribution' => $this->getCategoryDistribution(
-                        $user,
-                        $startDate,
-                        $endDate,
-                        $walletIds,
-                        $defaultCurrency
-                    ),
-                    'period_summary' => $this->getPeriodSummary(
-                        $user,
-                        $startDate,
-                        $endDate,
-                        $walletIds,
-                        $defaultCurrency
-                    ),
-                    'charts' => [
-                        'party_spending' => $this->getPartyData(
-                            $user,
-                            $startDate,
-                            $endDate,
-                            $walletIds,
-                            'expense',
-                            $defaultCurrency
-                        ),
-                        'party_income' => $this->getPartyData(
-                            $user,
-                            $startDate,
-                            $endDate,
-                            $walletIds,
-                            'income',
-                            $defaultCurrency
-                        ),
-                        'category_spending' => $this->getCategorySpendingData(
-                            $user,
-                            $startDate,
-                            $endDate,
-                            $walletIds,
-                            'expense',
-                            $defaultCurrency
-                        ),
-                        'income_sources' => $this->getCategorySpendingData(
-                            $user,
-                            $startDate,
-                            $endDate,
-                            $walletIds,
-                            'income',
-                            $defaultCurrency
-                        ),
-                        'monthly_cash_flow' => $this->getMonthlyCashFlowData(
-                            $user,
-                            $startDate,
-                            $endDate,
-                            $walletIds,
-                            $defaultCurrency
-                        ),
-                        'expense_by_wallet' => $this->getExpenseByWalletData(
-                            $user,
-                            $startDate,
-                            $endDate,
-                            $walletIds,
-                            $defaultCurrency
-                        ),
-                    ],
-                ];
-            }
+            )
         );
 
         return response()->json(['data' => $data]);
     }
 
-    /**
-     * Generate a deterministic cache key for stats based on request parameters.
-     *
-     * @param  int  $userId  The authenticated user's ID
-     * @param  Carbon  $startDate  Start of the date range
-     * @param  Carbon  $endDate  End of the date range
-     * @param  array  $walletIds  Filter by specific wallet IDs (empty for all)
-     * @param  string  $period  Grouping period (day, week, month, year)
-     * @return string The generated cache key
-     */
-    private function generateCacheKey(
-        int $userId,
-        Carbon $startDate,
-        Carbon $endDate,
-        array $walletIds,
-        string $period
-    ): string {
-        $version = Cache::get('stats:user:' . $userId . ':version', 1);
-
-        $params = [
-            'start' => $startDate->toDateString(),
-            'end' => $endDate->toDateString(),
-            'wallets' => implode(',', $walletIds),
-            'period' => $period,
-            'v' => $version,
-        ];
-
-        return 'stats:user:' . $userId . ':' . md5(json_encode($params));
-    }
-
-    /**
-     * Invalidate all stats cache for a user.
-     *
-     * For Redis, deletes all matching keys. For file/database cache,
-     * increments a version key to invalidate existing cache entries.
-     *
-     * @param  int  $userId  The user ID whose cache should be invalidated
-     */
-    public static function invalidateUserCache(int $userId): void
+    private function resolveDateRange(Request $request): array
     {
-        $pattern = 'stats:user:' . $userId . ':*';
+        $endDate = Carbon::now()->endOfDay();
+        $startDate = Carbon::now()->subDays(30)->startOfDay();
 
-        if (config('cache.default') === 'redis') {
-            $redis = Cache::getRedis();
-            $prefix = config('cache.prefix', 'laravel_cache');
-            $keys = $redis->keys("{$prefix}:{$pattern}");
-            foreach ($keys as $key) {
-                $redis->del($key);
-            }
+        if ($request->has('preset')) {
+            $endDate = Carbon::now()->endOfDay();
+
+            $startDate = match ($request->input('preset')) {
+                'all_time' => Carbon::parse('2000-01-01')->startOfDay(),
+                'current_week' => Carbon::now()->startOfWeek()->startOfDay(),
+                'current_month' => Carbon::now()->startOfMonth()->startOfDay(),
+                'last_3_months' => Carbon::now()->subMonths(3)->startOfDay(),
+                default => Carbon::now()->subDays(30)->startOfDay(),
+            };
         } else {
-            // For file/database cache, we can't easily pattern-match
-            // Instead, we'll use a version key approach
-            Cache::increment('stats:user:' . $userId . ':version');
+            if ($request->has('start_date')) {
+                $startDate = Carbon::parse($request->input('start_date'))->startOfDay();
+            }
+            if ($request->has('end_date')) {
+                $endDate = Carbon::parse($request->input('end_date'))->endOfDay();
+            }
         }
+
+        return [$startDate, $endDate];
     }
 
     /**
-     * Calculate total balance across user's wallets converted to target currency.
+     * @return array|JsonResponse
      */
-    private function getTotalBalance($user, array $walletIds = [], string $targetCurrency = 'USD'): float
+    private function resolveWalletIds(Request $request, $user): array|JsonResponse
     {
-        $query = Wallet::where('user_id', $user->id);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('id', $walletIds);
+        if (! $request->has('wallet_ids')) {
+            return [];
         }
 
-        $wallets = $query->get(['balance', 'currency']);
-        $total = 0.0;
-
-        foreach ($wallets as $wallet) {
-            $total += $this->convertCurrency(
-                (float) $wallet->balance,
-                $wallet->currency,
-                $targetCurrency,
-                $user
-            );
-        }
-
-        return $total;
-    }
-
-    /**
-     * Calculate user's net worth converted to target currency.
-     */
-    private function getNetWorth($user, array $walletIds = [], string $targetCurrency = 'USD'): float
-    {
-        return $this->getTotalBalance($user, $walletIds, $targetCurrency);
-    }
-
-    /**
-     * Calculate average monthly amount for a transaction type with currency conversion.
-     */
-    private function getAverageMonthlyAmount(
-        $user,
-        string $type,
-        array $walletIds,
-        Carbon $startDate,
-        Carbon $endDate,
-        string $targetCurrency = 'USD'
-    ): float {
-        $query = Transaction::join('wallets', 'transactions.wallet_id', '=', 'wallets.id')
-            ->where('transactions.user_id', $user->id)
-            ->where('transactions.type', $type)
-            ->nonTransfer()
-            ->whereBetween('transactions.datetime', [$startDate, $endDate])
-            ->select('transactions.amount', 'transactions.datetime', 'wallets.currency');
-
-        if (! empty($walletIds)) {
-            $query->whereIn('transactions.wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $monthlyTotals = [];
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->currency,
-                $targetCurrency,
-                $user
-            );
-
-            $monthKey = Carbon::parse($transaction->datetime)->format('Y-m');
-            $monthlyTotals[$monthKey] = ($monthlyTotals[$monthKey] ?? 0) + $converted;
-        }
-
-        if (empty($monthlyTotals)) {
-            return 0.0;
-        }
-
-        return array_sum($monthlyTotals) / count($monthlyTotals);
-    }
-
-    /**
-     * Get period-over-period comparisons with currency conversion.
-     */
-    private function getComparisons(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $daysInPeriod = $startDate->diffInDays($endDate);
-        $previousPeriodStart = $startDate->copy()->subDays($daysInPeriod);
-        $previousPeriodEnd = $startDate->copy()->subDay();
-
-        $currentPeriodTotals = $this->getPeriodTotals($user, $startDate, $endDate, $walletIds, $targetCurrency);
-        $previousPeriodTotals = $this->getPeriodTotals(
-            $user,
-            $previousPeriodStart,
-            $previousPeriodEnd,
-            $walletIds,
-            $targetCurrency
-        );
-
-        $calculateChange = function ($current, $previous) {
-            if ($previous == 0) {
-                return $current > 0 ? 100.0 : 0.0;
-            }
-
-            return (($current - $previous) / abs($previous)) * 100;
-        };
-
-        return [
-            'previous_period' => [
-                'income_change_percent' => $calculateChange(
-                    $currentPeriodTotals['income'],
-                    $previousPeriodTotals['income']
-                ),
-                'expense_change_percent' => $calculateChange(
-                    $currentPeriodTotals['expense'],
-                    $previousPeriodTotals['expense']
-                ),
-                'savings_rate_change' => $calculateChange(
-                    $currentPeriodTotals['savings_rate'],
-                    $previousPeriodTotals['savings_rate']
-                ),
-            ],
-        ];
-    }
-
-    /**
-     * Calculate income, expense totals for a period with currency conversion.
-     */
-    private function getPeriodTotals(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::join('wallets', 'transactions.wallet_id', '=', 'wallets.id')
-            ->where('transactions.user_id', $user->id)
-            ->nonTransfer()
-            ->whereBetween('transactions.datetime', [$startDate, $endDate])
-            ->select('transactions.amount', 'transactions.type', 'wallets.currency');
-
-        if (! empty($walletIds)) {
-            $query->whereIn('transactions.wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $income = 0.0;
-        $expense = 0.0;
-
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->currency,
-                $targetCurrency,
-                $user
-            );
-
-            if ($transaction->type === 'income') {
-                $income += $converted;
-            } else {
-                $expense += $converted;
-            }
-        }
-
-        $savingsRate = $income > 0 ? (($income - $expense) / $income) * 100 : 0;
-
-        return [
-            'income' => $income,
-            'expense' => $expense,
-            'savings_rate' => $savingsRate,
-        ];
-    }
-
-    /**
-     * Get top categories with currency conversion.
-     */
-    private function getTopCategories(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::with(['categories', 'wallet'])
-            ->where('user_id', $user->id)
-            ->nonTransfer()
-            ->whereBetween('datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $incomeTransactions = $transactions->where('type', 'income');
-        $expenseTransactions = $transactions->where('type', 'expense');
-
-        return [
-            'income' => $this->groupTransactionsByCategory($incomeTransactions, $user, $targetCurrency, 5),
-            'expenses' => $this->groupTransactionsByCategory($expenseTransactions, $user, $targetCurrency, 5),
-        ];
-    }
-
-    /**
-     * Get largest transactions with currency conversion.
-     */
-    private function getLargestTransactions(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::with(['categories', 'wallet'])
-            ->where('user_id', $user->id)
-            ->nonTransfer()
-            ->whereBetween('datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $findLargest = function ($type) use ($transactions, $user, $targetCurrency) {
-            $filtered = $transactions->where('type', $type);
-            if ($filtered->isEmpty()) {
-                return null;
-            }
-
-            $largest = null;
-            $largestAmount = 0;
-
-            foreach ($filtered as $transaction) {
-                $converted = $this->convertCurrency(
-                    (float) $transaction->amount,
-                    $transaction->wallet->currency ?? 'USD',
-                    $targetCurrency,
-                    $user
-                );
-
-                if ($converted > $largestAmount) {
-                    $largestAmount = $converted;
-                    $largest = [
-                        'amount' => $converted,
-                        'description' => $transaction->description,
-                        'date' => $transaction->datetime ? Carbon::parse($transaction->datetime)->toDateString() : null,
-                        'category' => $transaction->categories->first()?->name,
-                    ];
-                }
-            }
-
-            return $largest;
-        };
-
-        return [
-            'income' => $findLargest('income'),
-            'expense' => $findLargest('expense'),
-        ];
-    }
-
-    /**
-     * Get spending trends with currency conversion.
-     */
-    private function getSpendingTrends(
-        $user,
-        $startDate,
-        $endDate,
-        string $period = 'month',
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $currentPeriodData = $this->getTrendsForPeriod(
-            $user,
-            $startDate,
-            $endDate,
-            $period,
-            $walletIds,
-            $targetCurrency
-        );
-
-        $daysInPeriod = $startDate->diffInDays($endDate);
-        $previousStartDate = $startDate->copy()->subDays($daysInPeriod + 1);
-        $previousEndDate = $startDate->copy()->subDay();
-
-        $previousPeriodData = $this->getTrendsForPeriod(
-            $user,
-            $previousStartDate,
-            $previousEndDate,
-            $period,
-            $walletIds,
-            $targetCurrency
-        );
-
-        return [
-            'current_period' => $currentPeriodData,
-            'previous_period' => $previousPeriodData,
-        ];
-    }
-
-    /**
-     * Get expense trends with currency conversion.
-     */
-    private function getTrendsForPeriod(
-        $user,
-        $startDate,
-        $endDate,
-        $period,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::join('wallets', 'transactions.wallet_id', '=', 'wallets.id')
-            ->where('transactions.user_id', $user->id)
-            ->nonTransfer()
-            ->whereBetween('transactions.datetime', [$startDate, $endDate])
-            ->where('transactions.type', 'expense')
-            ->select('transactions.amount', 'transactions.datetime', 'wallets.currency');
-
-        if (! empty($walletIds)) {
-            $query->whereIn('transactions.wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $groupedData = [];
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->currency,
-                $targetCurrency,
-                $user
-            );
-
-            $key = $period === 'month'
-                ? Carbon::parse($transaction->datetime)->format('Y-m-01')
-                : Carbon::parse($transaction->datetime)->format('Y-m-d');
-
-            $groupedData[$key] = ($groupedData[$key] ?? 0) + $converted;
-        }
-
-        ksort($groupedData);
-
-        return collect($groupedData)->map(function ($amount, $date) {
-            return [
-                'date' => $date,
-                'amount' => (float) $amount,
-            ];
-        })->values()->toArray();
-    }
-
-    /**
-     * Get expense distribution by classification with currency conversion.
-     */
-    private function getCategoryDistribution(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::with(['categories', 'wallet'])
-            ->where('user_id', $user->id)
-            ->nonTransfer()
-            ->where('type', 'expense')
-            ->whereBetween('datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $distribution = [
-            'essential' => 0,
-            'non_essential' => 0,
-            'savings' => 0,
-            'investments' => 0,
-        ];
-
-        $total = 0;
-
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->wallet->currency ?? 'USD',
-                $targetCurrency,
-                $user
-            );
-
-            $total += $converted;
-
-            $category = $transaction->categories->first();
-            $classification = $this->classifyCategory($category);
-            $distribution[$classification] += $converted;
-        }
-
-        if ($total > 0) {
-            foreach ($distribution as $key => $amount) {
-                $distribution[$key] = ($amount / $total) * 100;
-            }
-        }
-
-        return $distribution;
-    }
-
-    /**
-     * Classify a category based on its name using pattern matching.
-     *
-     * @param  mixed  $category  The category model or null
-     * @return string Classification: essential, savings, investments, or non_essential
-     */
-    private function classifyCategory($category): string
-    {
-        if (! $category) {
-            return 'non_essential';
-        }
-
-        $name = strtolower($category->name ?? '');
-        $slug = strtolower($category->slug ?? '');
-        $text = $name . ' ' . $slug;
-
-        // Essential: housing, utilities, food, healthcare, transport, insurance
-        // @phpcs:ignore
-        if (preg_match('/\b(rent|mortgage|hous|utilit|electric|water|gas|grocer|food|meal|health|medic|pharma|doctor|hospital|insur|transport|fuel|petrol|diesel|commut|bus|train|taxi)\b/i', $text)) {
-            return 'essential';
-        }
-
-        // Savings
-        if (preg_match('/\b(saving|emergency|reserve|rainy.?day)\b/i', $text)) {
-            return 'savings';
-        }
-
-        // Investments
-        if (preg_match('/\b(invest|stock|bond|mutual|etf|crypto|retire|401k|ira|pension|dividend)\b/i', $text)) {
-            return 'investments';
-        }
-
-        return 'non_essential';
-    }
-
-    /**
-     * Get period summary with currency conversion.
-     */
-    private function getPeriodSummary(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $today = Carbon::now();
-        $daysRemaining = $today->diffInDays($endDate, false);
-        $daysElapsed = $startDate->diffInDays($today);
-        $totalDays = $startDate->diffInDays($endDate);
-
-        $totals = $this->getPeriodTotals($user, $startDate, $today, $walletIds, $targetCurrency);
-
-        $currentIncome = $totals['income'];
-        $currentExpenses = $totals['expense'];
-
-        $projectedIncome = $daysElapsed > 0
-            ? ($currentIncome / $daysElapsed) * $totalDays
-            : 0;
-
-        $projectedExpenses = $daysElapsed > 0
-            ? ($currentExpenses / $daysElapsed) * $totalDays
-            : 0;
-
-        return [
-            'start_date' => $startDate->toDateString(),
-            'end_date' => $endDate->toDateString(),
-            'days_remaining' => max(0, $daysRemaining),
-            'projected_income' => round($projectedIncome, 2),
-            'projected_expenses' => round($projectedExpenses, 2),
-        ];
-    }
-
-    /**
-     * Get party data with currency conversion.
-     */
-    private function getPartyData(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $type = 'expense',
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::with(['party', 'wallet'])
-            ->where('user_id', $user->id)
-            ->nonTransfer()
-            ->where('type', $type)
-            ->whereBetween('datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $partyData = [];
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->wallet->currency ?? 'USD',
-                $targetCurrency,
-                $user
-            );
-
-            $partyId = $transaction->party_id ?? 'uncategorized';
-
-            if (! isset($partyData[$partyId])) {
-                $partyData[$partyId] = [
-                    'id' => $partyId,
-                    'name' => $transaction->party ? $transaction->party->name : 'Uncategorized',
-                    'amount' => 0,
-                    'transaction_count' => 0,
-                    'percentage' => 0,
-                ];
-            }
-
-            $partyData[$partyId]['amount'] += $converted;
-            $partyData[$partyId]['transaction_count']++;
-        }
-
-        $totalAmount = array_sum(array_column($partyData, 'amount'));
-
-        return collect($partyData)
-            ->map(function ($item) use ($totalAmount) {
-                $item['percentage'] = $totalAmount > 0 ? ($item['amount'] / $totalAmount) * 100 : 0;
-
-                return $item;
-            })
-            ->sortByDesc('amount')
-            ->values()
-            ->toArray();
-    }
-
-    /**
-     * Get category spending data with currency conversion.
-     */
-    private function getCategorySpendingData(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds,
-        string $type = 'expense',
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::with(['categories', 'wallet'])
-            ->where('user_id', $user->id)
-            ->nonTransfer()
-            ->where('type', $type)
-            ->whereBetween('datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('wallet_id', $walletIds);
-        }
-
-        return $this->groupTransactionsByCategory($query->get(), $user, $targetCurrency);
-    }
-
-    /**
-     * Group transactions by category with currency conversion.
-     */
-    private function groupTransactionsByCategory(
-        $transactions,
-        $user,
-        string $targetCurrency,
-        ?int $limit = null
-    ): array {
-        $categoryData = [];
-
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->wallet->currency ?? 'USD',
-                $targetCurrency,
-                $user
-            );
-
-            $category = $transaction->categories->first();
-
-            if (! $category) {
-                $categoryId = 'uncategorized';
-                $categoryName = 'Uncategorized';
-            } else {
-                $categoryId = $category->id;
-                $categoryName = $category->name;
-            }
-
-            if (! isset($categoryData[$categoryId])) {
-                $categoryData[$categoryId] = [
-                    'id' => $categoryId,
-                    'name' => $categoryName,
-                    'amount' => 0,
-                    'transaction_count' => 0,
-                ];
-            }
-
-            $categoryData[$categoryId]['amount'] += $converted;
-            $categoryData[$categoryId]['transaction_count']++;
-        }
-
-        $totalAmount = array_sum(array_column($categoryData, 'amount'));
-
-        $result = collect($categoryData)
-            ->sortByDesc('amount')
-            ->when($limit, fn ($collection) => $collection->take($limit))
-            ->map(function ($item) use ($totalAmount) {
-                $item['amount'] = (float) $item['amount'];
-                $item['percentage'] = $totalAmount > 0 ? ($item['amount'] / $totalAmount) * 100 : 0;
-
-                return $item;
-            })
-            ->values()
+        $walletIds = array_filter(array_map('intval', explode(',', $request->input('wallet_ids'))));
+
+        $validWalletIds = Wallet::where('user_id', $user->id)
+            ->whereIn('id', $walletIds)
+            ->pluck('id')
             ->toArray();
 
-        return $result;
-    }
-
-    /**
-     * Get monthly cash flow data with currency conversion.
-     */
-    private function getMonthlyCashFlowData(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::join('wallets', 'transactions.wallet_id', '=', 'wallets.id')
-            ->where('transactions.user_id', $user->id)
-            ->nonTransfer()
-            ->whereBetween('transactions.datetime', [$startDate, $endDate])
-            ->select('transactions.amount', 'transactions.type', 'transactions.datetime', 'wallets.currency');
-
-        if (! empty($walletIds)) {
-            $query->whereIn('transactions.wallet_id', $walletIds);
+        $invalidWalletIds = array_diff($walletIds, $validWalletIds);
+        if (! empty($invalidWalletIds)) {
+            return response()->json([
+                'message' => __('One or more wallet IDs are invalid or do not belong to the user.'),
+                'invalid_wallet_ids' => array_values($invalidWalletIds),
+            ], 422);
         }
 
-        $transactions = $query->get();
-
-        $monthlyData = [];
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->currency,
-                $targetCurrency,
-                $user
-            );
-
-            $period = Carbon::parse($transaction->datetime)->format('Y-m');
-
-            if (! isset($monthlyData[$period])) {
-                $monthlyData[$period] = ['income' => 0, 'expense' => 0];
-            }
-
-            if ($transaction->type === 'income') {
-                $monthlyData[$period]['income'] += $converted;
-            } else {
-                $monthlyData[$period]['expense'] += $converted;
-            }
-        }
-
-        ksort($monthlyData);
-
-        return collect($monthlyData)->map(function ($data, $period) {
-            return [
-                'period' => $period,
-                'income' => (float) $data['income'],
-                'expense' => (float) $data['expense'],
-                'net' => (float) ($data['income'] - $data['expense']),
-            ];
-        })->values()->toArray();
-    }
-
-    /**
-     * Get expense by wallet with currency conversion.
-     */
-    private function getExpenseByWalletData(
-        $user,
-        $startDate,
-        $endDate,
-        array $walletIds = [],
-        string $targetCurrency = 'USD'
-    ): array {
-        $query = Transaction::with('wallet')
-            ->where('user_id', $user->id)
-            ->nonTransfer()
-            ->where('type', 'expense')
-            ->whereBetween('datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $query->whereIn('wallet_id', $walletIds);
-        }
-
-        $transactions = $query->get();
-
-        $walletData = [];
-        foreach ($transactions as $transaction) {
-            $converted = $this->convertCurrency(
-                (float) $transaction->amount,
-                $transaction->wallet->currency ?? 'USD',
-                $targetCurrency,
-                $user
-            );
-
-            $walletId = $transaction->wallet_id;
-
-            if (! isset($walletData[$walletId])) {
-                $walletData[$walletId] = [
-                    'wallet_id' => $walletId,
-                    'wallet_name' => $transaction->wallet->name ?? 'Unknown',
-                    'amount' => 0,
-                    'transaction_count' => 0,
-                ];
-            }
-
-            $walletData[$walletId]['amount'] += $converted;
-            $walletData[$walletId]['transaction_count']++;
-        }
-
-        $totalAmount = array_sum(array_column($walletData, 'amount'));
-
-        return collect($walletData)
-            ->map(function ($item) use ($totalAmount) {
-                $item['percentage'] = $totalAmount > 0 ? ($item['amount'] / $totalAmount) * 100 : 0;
-
-                return $item;
-            })
-            ->sortByDesc('amount')
-            ->values()
-            ->toArray();
-    }
-
-    /**
-     * Get activity metrics: counts, frequencies, busiest day, top party/category by frequency.
-     */
-    private function getActivityMetrics(
-        $user,
-        Carbon $startDate,
-        Carbon $endDate,
-        array $walletIds = []
-    ): array {
-        $baseQuery = Transaction::where('transactions.user_id', $user->id)
-            ->nonTransfer()
-            ->whereBetween('transactions.datetime', [$startDate, $endDate]);
-
-        if (! empty($walletIds)) {
-            $baseQuery->whereIn('transactions.wallet_id', $walletIds);
-        }
-
-        $stats = (clone $baseQuery)->selectRaw('
-            count(*) as total_count,
-            count(distinct party_id) as unique_parties,
-            sum(case when type = "income" then amount else 0 end) as total_income,
-            count(case when type = "income" then 1 end) as income_count,
-            sum(case when type = "expense" then amount else 0 end) as total_expenses,
-            count(case when type = "expense" then 1 end) as expense_count
-        ')->first();
-
-        $transactionCount = (int) $stats->total_count;
-        $periodDays = max(1, $startDate->diffInDays($endDate));
-
-        $avgIncome = $stats->income_count > 0 ? round($stats->total_income / $stats->income_count, 2) : 0;
-        $avgExpense = $stats->expense_count > 0 ? round($stats->total_expenses / $stats->expense_count, 2) : 0;
-
-        return [
-            'transaction_count' => $transactionCount,
-            'unique_parties' => (int) $stats->unique_parties,
-            'average_income_transaction' => $avgIncome,
-            'average_expense_transaction' => $avgExpense,
-            'frequency' => [
-                'per_day' => round($transactionCount / $periodDays, 2),
-                'per_week' => round($transactionCount / max(1, $periodDays / 7), 2),
-                'per_month' => round($transactionCount / max(1, $periodDays / 30), 2),
-            ],
-            'busiest_day' => $this->getBusiestDay($baseQuery),
-            'most_frequent_party' => $this->getMostFrequentParty($baseQuery),
-            'most_used_category' => $this->getMostUsedCategory($baseQuery),
-        ];
-    }
-
-    private function getBusiestDay($baseQuery): ?string
-    {
-        $row = (clone $baseQuery)
-            ->selectRaw('DAYNAME(datetime) as day_name, count(*) as cnt')
-            ->groupByRaw('DAYNAME(datetime)')
-            ->orderByDesc('cnt')
-            ->first();
-
-        return $row?->day_name;
-    }
-
-    private function getMostFrequentParty($baseQuery): ?array
-    {
-        $row = (clone $baseQuery)
-            ->join('parties', 'transactions.party_id', '=', 'parties.id')
-            ->selectRaw('parties.id, parties.name, count(*) as transaction_count')
-            ->groupBy('parties.id', 'parties.name')
-            ->orderByDesc('transaction_count')
-            ->first();
-
-        if (! $row) {
-            return null;
-        }
-
-        return [
-            'id' => $row->id,
-            'name' => $row->name,
-            'transaction_count' => (int) $row->transaction_count,
-        ];
-    }
-
-    private function getMostUsedCategory($baseQuery): ?array
-    {
-        $row = \DB::table('categorizables')
-            ->join('categories', 'categorizables.category_id', '=', 'categories.id')
-            ->where('categorizables.categorizable_type', Transaction::class)
-            ->whereIn('categorizables.categorizable_id', (clone $baseQuery)->select('transactions.id'))
-            ->selectRaw('categories.id, categories.name, count(*) as transaction_count')
-            ->groupBy('categories.id', 'categories.name')
-            ->orderByDesc('transaction_count')
-            ->first();
-
-        if (! $row) {
-            return null;
-        }
-
-        return [
-            'id' => $row->id,
-            'name' => $row->name,
-            'transaction_count' => (int) $row->transaction_count,
-        ];
+        return $walletIds;
     }
 }
